refactor: rewrite module cjsToEsm.cjs in ES6 and rename to esm.mjs
